Refactored register indexing with usize

+++ simulator_engine/src/main.rs
@@ -30,9 +30,9 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
         let opcode = instruction & 0x7f;
         let funct3 = (instruction >> 12) & 0x07;
         let funct7 = instruction >> 25;
-        let rd = (instruction >> 7) & 0x01f;
-        let rs1 = (instruction >> 15) & 0x01f;
-        let rs2 = (instruction >> 20) & 0x01f;
+        let rd = ((instruction >> 7) & 0x01f) as usize;
+        let rs1 = ((instruction >> 15) & 0x01f) as usize;
+        let rs2 = ((instruction >> 20) & 0x01f) as usize;
         let imm3112 = (instruction >> 12) << 12;
         let imm110 = instruction >> 20;
         let shamt = (instruction >> 20) & 0x01f;
@@ -40,12 +40,12 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
         match opcode {
             0x03 => match funct3 {
                 0x00 => {
-                    reg[rd as usize] = mem[(reg[rs1 as usize] + imm110) as usize] as i32;
+                    reg[rd] = mem[(reg[rs1] + imm110) as usize] as i32;
                     println!("LB x{}, x{}, {}", rd, rs1, imm110);
                 }
                 0x01 => {
-                    let short: [u8; 4] = [mem[(reg[rs1 as usize] + imm110) as usize], mem[((reg[rs1 as usize] + imm110) + 1) as usize], 0, 0];
-                    reg[rd as usize] = i32::from_be_bytes(short);
+                    let short: [u8; 4] = [mem[(reg[rs1] + imm110) as usize], mem[((reg[rs1] + imm110) + 1) as usize], 0, 0];
+                    reg[rd] = i32::from_be_bytes(short);
                     println!("LH x{}, x{}, {}", rd, rs1, imm110);
                 }
                 unimplemented => println!(
@@ -55,41 +55,41 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
             }
             0x13 => match funct3 {
                 0x00 => {
-                    reg[rd as usize] = reg[rs1 as usize] + imm110;
+                    reg[rd] = reg[rs1] + imm110;
                     println!("ADDI x{}, x{}, {}", rd, rs1, imm110);
                 }
                 0x01 => {
-                    reg[rd as usize] = reg[rs1 as usize] << shamt;
+                    reg[rd] = reg[rs1] << shamt;
                     println!("SLLI x{}, x{}, {}", rd, rs1, shamt);
                 }
                 0x02 => {
-                    if reg[rs1 as usize] < imm110 {
-                        reg[rd as usize] = 1;
+                    if reg[rs1] < imm110 {
+                        reg[rd] = 1;
                     } else {
-                        reg[rd as usize] = 0;
+                        reg[rd] = 0;
                     }
                     println!("SLTI x{}, x{}, {}", rd, rs1, imm110);
                 }
                 0x03 => {
-                    if (reg[rs1 as usize] as u32) < (imm110 as u32) {
-                        reg[rd as usize] = 1;
+                    if (reg[rs1] as u32) < (imm110 as u32) {
+                        reg[rd] = 1;
                     } else {
-                        reg[rd as usize] = 0;
+                        reg[rd] = 0;
                     }
                     println!("SLTIU x{}, x{}, {}", rd, rs1, imm110);
                 }
                 0x04 => {
-                    reg[rd as usize] = reg[rs1 as usize] ^ imm110;
+                    reg[rd] = reg[rs1] ^ imm110;
                     println!("XORI x{}, x{}, {}", rd, rs1, imm110);
                 }
                 // TODO:
                 0x05 => match funct7 {
                     0x00 => {
-                        reg[rd as usize] = ((reg[rs1 as usize] as u32) >> (shamt as u32)) as i32;
+                        reg[rd] = ((reg[rs1] as u32) >> (shamt as u32)) as i32;
                         println!("SRLI x{}, x{}, {}", rd, rs1, shamt);
                     }
                     0x20 => {
-                        reg[rd as usize] = reg[rs1 as usize] >> shamt;
+                        reg[rd] = reg[rs1] >> shamt;
                         println!("SRAI x{}, x{}, {}", rd, rs1, shamt);
                     }
                     unimplemented => println!(
@@ -98,11 +98,11 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                     ),
                 },
                 0x06 => {
-                    reg[rd as usize] = reg[rs1 as usize] | imm110;
+                    reg[rd] = reg[rs1] | imm110;
                     println!("ORI x{}, x{}, {}", rd, rs1, imm110);
                 }
                 0x07 => {
-                    reg[rd as usize] = reg[rs1 as usize] & imm110;
+                    reg[rd] = reg[rs1] & imm110;
                     println!("ANDI x{}, x{}, {}", rd, rs1, imm110);
                 }
                 unimplemented => println!(
@@ -111,17 +111,17 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                 ),
             },
             0x17 => {
-                reg[rd as usize] = pc as i32 + imm3112;
+                reg[rd] = pc as i32 + imm3112;
                 println!("AUIPC x{}, {}", rd, imm3112);
             }
             0x33 => match funct3 {
                 0x00 => match funct7 {
                     0x00 => {
-                        reg[rd as usize] = reg[rs1 as usize] + reg[rs2 as usize];
+                        reg[rd] = reg[rs1] + reg[rs2];
                         println!("ADD x{}, x{}, x{}", rd, rs1, rs2);
                     }
                     0x20 => {
-                        reg[rd as usize] = reg[rs1 as usize] - reg[rs2 as usize];
+                        reg[rd] = reg[rs1] - reg[rs2];
                         println!("SUB x{}, x{}, x{}", rd, rs1, rs2);
                     }
                     unimplemented => println!(
@@ -130,36 +130,36 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                     ),
                 },
                 0x01 => {
-                    reg[rd as usize] = reg[rs1 as usize] << reg[rs2 as usize];
+                    reg[rd] = reg[rs1] << reg[rs2];
                     println!("SLL x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 0x02 => {
-                    if reg[rs1 as usize] < reg[rs2 as usize] {
-                        reg[rd as usize] = 1;
+                    if reg[rs1] < reg[rs2] {
+                        reg[rd] = 1;
                     } else {
-                        reg[rd as usize] = 0;
+                        reg[rd] = 0;
                     }
                     println!("SLT x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 0x03 => {
-                    if (reg[rs1 as usize] as u32) < (reg[rs2 as usize] as u32) {
-                        reg[rd as usize] = 1;
+                    if (reg[rs1] as u32) < (reg[rs2] as u32) {
+                        reg[rd] = 1;
                     } else {
-                        reg[rd as usize] = 0;
+                        reg[rd] = 0;
                     }
                     println!("SLTIU x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 0x04 => {
-                    reg[rd as usize] = reg[rs1 as usize] ^ reg[rs2 as usize];
+                    reg[rd] = reg[rs1] ^ reg[rs2];
                     println!("XOR x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 0x05 => match funct7 {
                     0x00 => {
-                        reg[rd as usize] = ((reg[rs1 as usize] as u32) >> (reg[rs2 as usize] as u32)) as i32;
+                        reg[rd] = ((reg[rs1] as u32) >> (reg[rs2] as u32)) as i32;
                         println!("SRL x{}, x{}, x{}", rd, rs1, rs2);
                     }
                     0x20 => {
-                        reg[rd as usize] = reg[rs1 as usize] >>  reg[rs2 as usize];
+                        reg[rd] = reg[rs1] >>  reg[rs2];
                         println!("SRA x{}, x{}, x{}", rd, rs1, rs2);
                     }
                     unimplemented => println!(
@@ -168,11 +168,11 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                     ),
                 }
                 0x06 => {
-                    reg[rd as usize] = reg[rs1 as usize] | reg[rs2 as usize];
+                    reg[rd] = reg[rs1] | reg[rs2];
                     println!("OR x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 0x07 => {
-                    reg[rd as usize] = reg[rs1 as usize] & reg[rs2 as usize];
+                    reg[rd] = reg[rs1] & reg[rs2];
                     println!("AND x{}, x{}, x{}", rd, rs1, rs2);
                 }
                 unimplemented => println!(
@@ -181,42 +181,42 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                 ),
             },
             0x37 => {
-                reg[rd as usize] = imm3112;
+                reg[rd] = imm3112;
                 println!("LUI x{}, {}", rd, imm3112);
             }
             0x63 => match funct3 {
                 0x00 => {
-                    if reg[rs1 as usize] == reg[rs2 as usize] {
+                    if reg[rs1] == reg[rs2] {
                         pc += sb_format(&instruction);
                     };
                     println!("BEQ x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
                 }
                 0x01 => {
-                    if reg[rs1 as usize] != reg[rs2 as usize] {
+                    if reg[rs1] != reg[rs2] {
                         pc += sb_format(&instruction);
                     };
                     println!("BNE x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
                 }
                 0x04 => {
-                    if reg[rs1 as usize] < reg[rs2 as usize] {
+                    if reg[rs1] < reg[rs2] {
                         pc += sb_format(&instruction);
                     };
                     println!("BLT x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
                 }
                 0x05 => {
-                    if reg[rs1 as usize] >= reg[rs2 as usize] {
+                    if reg[rs1] >= reg[rs2] {
                         pc += sb_format(&instruction);
                     };
                     println!("BGE x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
                 }
                 0x06 => {
-                    if (reg[rs1 as usize] as u32) < (reg[rs2 as usize] as u32) {
+                    if (reg[rs1] as u32) < (reg[rs2] as u32) {
                         pc += sb_format(&instruction);
                     };
                     println!("BLTU x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
                 }
                 0x07 => {
-                    if (reg[rs1 as usize] as u32) >= (reg[rs2 as usize] as u32) {
+                    if (reg[rs1] as u32) >= (reg[rs2] as u32) {
                         pc += sb_format(&instruction);
                     };
                     println!("BGEU x{}, x{}, {}", rs1, rs2, sb_format(&instruction));
@@ -228,8 +228,8 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
             },
             0x67 => match funct3 {
                 0x00 => {
-                    reg[rd as usize] = pc as i32 + 4;
-                    pc = (reg[rs1 as usize] + imm110) as usize;
+                    reg[rd] = pc as i32 + 4;
+                    pc = (reg[rs1] + imm110) as usize;
                     println!("JALR x{}, x{}, {}", rd, rs1, imm110);
                 }
                 unimplemented => println!(
@@ -238,7 +238,7 @@ fn simulate(mut reg: [i32; 32], mut mem: [u8; 1048576], program_len: &usize) -> 
                 ),
             },
             0x6F => {
-                reg[rd as usize] = (pc + 4) as i32;
+                reg[rd] = (pc + 4) as i32;
                 pc = pc + uj_format(&instruction);
                 println!("JAL x{}, {}", rd, uj_format(&instruction));
             }
